Split Knowledge.Faq into two modules - one for FAQ and one for questions
